+++ cr_stripe_advanced_fees/models/__init__.py
@@ -5,3 +5,4 @@ from . import payment_provider
 from . import payment_transaction
 from . import payment_method_fees
 from . import controller
+from . import mail_followers